requirements.txt
argcomplete==1.10.0
Automat==20.2.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.0
certifi==2020.12.5
chardet==3.0.4
cobble==0.1.3
constantly==15.1.0
cssselect==1.1.0
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
html-text==0.5.2
html2text==2020.1.16
IMAPClient==2.1.0
incremental==17.5.0
ipykernel==5.5.0
jedi==0.18.0
jupyter-client==6.1.11
mammoth==1.4.15
MarkupSafe==1.1.1
mistune==0.8.4

nbconvert==5.6.1
olefile==0.46
pdfminer.six==20181108
Pillow==8.1.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome==3.10.1
PyDispatcher==2.0.5
PyHamcrest==2.0.2
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2021.1
pywin32==300
pyzmq==22.0.3
queuelib==1.5.0
service-identity==18.1.0
sortedcontainers==2.3.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tzlocal==1.5.1
webencodings==0.5.1
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.3.7
